1.建立异常拦截器，用于处理控制器中，业务层、工具类等抛出的异常信息，减少控制器的异常处理代码 2.配合异常处理拦截器，优化控制层的异常处理代码，减少相关操作，并测试成功 3.对管理账户删除进行显示，现在不在能自己删除自己了（包括单个删除和批量删除）
待完成：
1.对异常类进行划分，除了web异常，还应该增加权限异常等（后续根据实际要求确认）
2.redis跨域登陆，增加登陆的性能
3.产品分类模块编写，初步计划通过表格直接添加产品分类

+++ src/main/java/com/lcvc/ebuy_springboot/service/impl/AdminServiceImpl.java
@@ -27,9 +27,9 @@ public class AdminServiceImpl implements AdminService {
     public boolean login(String username, String password) throws MyFormException{
         boolean judge=false;
         if(StringUtils.isEmpty(username)){
-            throw new MyFormException("账户名不能为空");
+            throw new MyFormException("登陆失败：账户名不能为空");
         }else  if(StringUtils.isEmpty(password)){
-            throw new MyFormException("密码不能为空");
+            throw new MyFormException("登陆失败：密码不能为空");
         }else{
             if(adminDao.login(username, SHA.getResult(password))==1){
                 judge=true;
@@ -53,12 +53,22 @@ public class AdminServiceImpl implements AdminService {
 
 
     @Override
-    public void deleteAdmin(Integer id){
-        adminDao.delete(id);
+    public void deleteAdmin(Admin admin,Integer id) throws MyFormException{
+        if(admin.getId()==id.intValue()) {//如果登录账户的id与被删除账户的id一致
+            throw new MyFormException("删除失败：不允许删除自己的账户");
+        }else{
+            adminDao.delete(id);
+        }
     }
 
     @Override
-    public void deleteAdmins(Integer[] ids) {
+    public void deleteAdmins(Admin admin,Integer[] ids)  throws MyFormException{
+        //先进行验证
+        for(Integer id:ids){
+            if(admin.getId()==id.intValue()) {//如果登录账户的id与被删除账户的id一致
+                throw new MyFormException("删除失败：不允许删除自己的账户");
+            }
+        }
         adminDao.deletes(ids);
     }
 
@@ -66,20 +76,20 @@ public class AdminServiceImpl implements AdminService {
     public void addAdmin(Admin admin) throws MyFormException {
         if(admin!=null){
             if(admin.getUsername().equals("")){
-                throw new MyFormException("添加失败：账户名不能为空");
+                throw new MyFormException("账户添加失败：账户名不能为空");
             }
             if(admin.getName().equals("")){
-                throw new MyFormException("添加失败：姓名不能为空");
+                throw new MyFormException("账户添加失败：姓名不能为空");
             }
             if(adminDao.countUsername(admin.getUsername())==0){
                 admin.setPassword(SHA.getResult("123456"));
                 admin.setCreateTime(Calendar.getInstance().getTime());//获取当前时间为创建时间
                 adminDao.save(admin);
             }else{
-                throw new MyFormException("添加失败：账户名重名");
+                throw new MyFormException("账户添加失败：账户名重名");
             }
         }else{
-            throw new MyFormException("添加失败：表单数据不能为空");
+            throw new MyFormException("账户添加失败：表单数据不能为空");
         }
     }
 
@@ -105,9 +115,9 @@ public class AdminServiceImpl implements AdminService {
     public boolean updateAdmin(Admin admin) throws MyFormException {
         boolean status = false;// 默认编辑失败
         if(admin.getUsername().length()==0){
-            throw new MyFormException("账户名不能为空");
+            throw new MyFormException("账户编辑失败：账户名不能为空");
         }else if(adminDao.countOtherUsername(admin.getUsername(),admin.getId())>0) {//如果账户名重名
-            throw new MyFormException("和其他管理账户的账户名重名");
+            throw new MyFormException("账户编辑失败：和其他管理账户的账户名重名");
         }else{
             int i = adminDao.update(admin);// 更改了多少条记录
             // 编写代码，判断是否编辑成功
